backend/requirements.txt
# Environment
python-dotenv

# Django
Django

# DRF
django-filter
drf-flex-fields
drf-access-policy
drf-nested-routers
drf-writable-nested
djangorestframework
djangorestframework-simplejwt

# Databases
psycopg2-binary

# AWS
boto3

# AI
openai
replicate

# Image processing
opencv-python
Pillow

# HTTP-Requests
requests